+++ rollup.config.js
@@ -1,44 +1,22 @@
 import commonjs from "rollup-plugin-commonjs";
 import resolve from "rollup-plugin-node-resolve";
 import typescript from "rollup-plugin-typescript";
-import jscc from "rollup-plugin-jscc";
 import pkg from "./package.json";
-
-const plugins = [
-	typescript({
-		typescript: require("typescript"),
-	}),
-	resolve(),
-	commonjs(),
-];
 
 export default [
 	{
-		entry: "src/index.ts",
-		moduleName: "deckstrings",
-		dest: pkg.main,
-		format: "cjs",
+		input: "src/index.ts",
+		name: "deckstrings",
+		output: {
+			file: pkg.main,
+			format: "umd",
+		},
 		plugins: [
-			jscc({
-				values: {
-					_PLATFORM: "node",
-				},
-				extensions: [".js", ".ts"],
+			typescript({
+				typescript: require("typescript"),
 			}),
-		].concat(plugins),
-	},
-	{
-		entry: "src/index.ts",
-		moduleName: "deckstrings",
-		dest: pkg.browser,
-		format: "umd",
-		plugins: [
-			jscc({
-				values: {
-					_PLATFORM: "web",
-				},
-				extensions: [".js", ".ts"],
-			}),
-		].concat(plugins),
+			resolve(),
+			commonjs(),
+		],
 	},
 ];