+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add Typescript typings
+
 ### Changed
 - Switch to a single UMD build
 - Always return canonical deckstrings and deck definitions