+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.0.0] - 2018-02-09
 ### Added
 - Added browser bundles
 - Added ES module bundles
@@ -45,7 +47,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fix a decoding bug
 
-[Unreleased]: https://github.com/HearthSim/npm-deckstrings/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/HearthSim/npm-deckstrings/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/HearthSim/npm-deckstrings/compare/1.1.0...v2.0.0
 [1.1.0]: https://github.com/HearthSim/npm-deckstrings/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/HearthSim/npm-deckstrings/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/HearthSim/npm-deckstrings/compare/0.2.0...1.0.0